Fixed bugs in Hdd Manag & Started coding Mainframe
Forgot to code the sync for the HardDriveList Edit method so
edited harddrives weren't properly saved.

Added a toolbar to the mainframe with some buttons and a
combobox to allow the user to select the HDD category to
analyze.

Added the ability to associate a category with a Hdd thus
facilitating the whole process of category management and
path analyzing.