Implemented the FileProvider
Finished the file provider and adapted the Movie, Category and HardDrive
class to use the provider.

Each harddrive has a provider and all the categories and movies of that
harddrive relay the Load/Save operations to the HardDrive object which
then relays the operations to the provider. This prevents having loads
of provider references and having to change them when the harddrive one
changes.