Finished DatabaseProvider and started DBFileSync
Some problems arised with the implementation of DBFileSync that forced
me to change the structure and chain of operations regarding movie
loading and saving.

Now all movie info is easily obtained by means of a info dictionary.
This dictionary is created on movie info load and the movie info on hte
object is set by iterating over all the keys of the dictionary.
A similar thing happens on save. A info dictionary is obtained from the
movie object containing all its current values. Data to save is obtained
from this dict object.
This results in a more abstract loading and saving of information, allowing
easy and simple adding/removing of movie info fields.