Bugfixes. Started work on TMDB integration.
IMDB integration was being a pain in the ass. IMDB blocked all requests
after 20 or so queries so had to look for an alternative. Found TMDB and
since it provides an open API I figure it must be a little more liberal
about data gathering than IMDB.

Included in the project the 'themoviedb' script by doganaydin
(https://github.com/doganaydin/themoviedb) and that will be the base of
the TMDB integration.

I expect I should have everything working tomorrow.