InfoProvider is now TMDB.
Finished the integration of TMDB and MHDD Organizer. Info is gathered
much faster now!

Added a custom search to the TMDB Search Result dialog. The search
algorithm of TMDB isn't as flexible as IMDB's and, as such, there's a
higher chance that it wont recognize the movie in a first search so
there should be an easy way to change the search query (and there is now
:D)