Replaces plot with overview. Fix sync checking.
To reflect the movie structure in TMDB and to allow more freedom to the
text in the previous Plot field, it is now an Overview field.

Fixed a bug in sync that resulted in saving to DB or files each time the
movie list was loaded even if movies were already synced.